filename, cooldown, and timer update
Filenames (.png and .json) match the map name. Map name/author always defaults to 'Untitled' and 'Anonymous'. Portal cooldown and button timer defaults to 0, right click on individual portal or button to change. Or right click on the portal/button icon in the palette on the left side of the screen to modify all portal/button cooldowns/timers.